Update my-bind9.xml
The comment about you must change is false.  Thats what [IP]:[PORT:] is all about.  It takes the container port you've specified there and changes it to whatever host port the user specifies

And fill out the Targets on the ports.  With them filled out, the user doesn't have to actually specify 53